+++ src/hooks/useAsyncAction/index.ts
@@ -1,1 +1,2 @@
+export type { UseAsyncActionOptions } from "./useAsyncAction";
 export { useAsyncAction } from "./useAsyncAction";